+++ components/apimgt/org.wso2.carbon.apimgt.core/src/main/java/org/wso2/carbon/apimgt/core/configuration/models/KeyMgtConfigurations.java
@@ -28,7 +28,7 @@ import org.wso2.carbon.config.annotation.Element;
 public class KeyMgtConfigurations {
 
     @Element(description = "Key Manager Implementation class")
-    private String keyManagerImplClass = "org.wso2.carbon.apimgt.core.impl.LightWeightKeyManagerImpl";
+    private String keyManagerImplClass = "org.wso2.carbon.apimgt.core.impl.DefaultKeyManagerImpl";
     @Element(description = "DCR Endpoint URL")
     private String dcrEndpoint = "https://localhost:9282/api/identity/oauth2/dcr/v1.0/register";
     @Element(description = "Token Endpoint URL")
