+++ components/apimgt/org.wso2.carbon.apimgt.core/src/main/java/org/wso2/carbon/apimgt/core/impl/DefaultIdentityProviderImpl.java
@@ -52,7 +52,7 @@ import java.util.List;
 /**
  * API Manager default implementation of {@link IdentityProvider}
  */
-public class DefaultIdentityProviderImpl extends LightWeightKeyManagerImpl implements IdentityProvider {
+public class DefaultIdentityProviderImpl extends DefaultKeyManagerImpl implements IdentityProvider {
 
     private static final Logger log = LoggerFactory.getLogger(DefaultIdentityProviderImpl.class);
